Move Output to new architecture and update Print.
Change the Output plugins so that all parameters are worked out
and sanitised and set within the Output class (or more accurately
the relevant subclass). This is instead of using the AirPi.py method
define_plugin_params(), which was never really the right place to
do it. Only outputs/print.py has been updated to the new architecture
in this commit; the rest are still to be converted.

No changes have been made to the Sensor or Notification plugin
architectures.